compile ts file inside buildFile
